Fixed type for parameter margin in VisuBump::showWrappingShapes(..)

+++ src/main/java/fr/unice/polytech/si3/qgl/stormbreakers/visuals/bumps/VisuBump.java
@@ -64,7 +64,7 @@ public class VisuBump {
      * @param indexes if null shows all wrapping shapes
      *        {@code new int[] {2,4,6}} will show shapes of indexes 2,4,6
      */
-    private static void showWrappingShapes(List<Recif> reefs, int[] indexes, int margin, Displayer displayer) {
+    private static void showWrappingShapes(List<Recif> reefs, int[] indexes, double margin, Displayer displayer) {
         if (indexes != null) {
             for (int idx : indexes) {
                 showWrappingShape(reefs.get(idx), margin, displayer);
